Remove by list

+++ kompas/ips.remove/remove-ips.ps1
@@ -1,7 +1,15 @@
-$who = 'uxm00281'
+$hosts = Get-Content (Join-Path (Split-Path $PSCommandPath -Parent) hosts.txt)
 
-Invoke-Command -ComputerName $who -ScriptBlock {
-  Stop-Service IPSUpdater
-  Remove-Item HKLM:\SOFTWARE\Intermech -Force -Recurse
-  Remove-Item C:\IPS -Force -Recurse
+foreach ($who in $hosts) {
+
+  if (!(Test-Connection $who -Quiet -Count 1)) {
+    "Cannot connect to: $who"
+    continue
+  }
+  Invoke-Command -ComputerName $who -ScriptBlock {
+    Stop-Service IPSUpdater
+    Remove-Item HKLM:\SOFTWARE\Intermech -Force -Recurse
+    Remove-Item C:\IPS -Force -Recurse
+    Remove-Item "C:\IPS Vault" -Force -Recurse
+  }
 }
